+++ src/frontend/web_application/test/functional/features/00_home-spec.js
@@ -20,7 +20,7 @@ describe('Home', () => {
     it('Log In', () => {
       userUtil.signin();
       browser.get('/');
-      expect(element(by.css('.m-application-switcher .m-navbar-item__content')).getText()).toContain('Discussions');
+      expect(element(by.css('.m-application-switcher .m-navbar-item__content')).getText()).toContain('Messages');
     });
 
     it('Requires authentication', () => {