+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- keep scroll position between tabs
 - do not force phone nnumber normalization, accept everything and try to normalize
 - permit to set contact title on user input, do not compute it strictly
 - Reset password API