+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - do not force phone nnumber normalization, accept everything and try to normalize
 - permit to set contact title on user input, do not compute it strictly
 
+### Changed
+
+- The name of the Timeline tab is now "Messages" instead of "Discussions"
+
 ### Fixed
 
 - Disable send message when already sent and add visual feedback (spinner)