+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+- Do not close the dropdown when receiving new suggestions
 - Disable send message when already sent and add visual feedback (spinner)
 - Disable import contact button on uploading and add a spinner
 - Save updated password strength after password modification